Do not relay on Linux cooked-mode capture (SLL)
Linux cooked-mode capture (SLL) is the pseudo-protocol used by libpcap
on Linux when capturing on the "any" interface.  If you don't use
"any", you won't get it.  That means that if you capture on a one CAN
device, the dissectors don't work.

It seems SLL is only used for heuristic determination of CSP packets
and for the key, to which sll_pkt is always "1:".

Remove sll.pkttype and sll.ltype so that the dissector works on
captures without cooked-mode capture.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>